[carddav] Improve delta detection. Contributes to MER#941
This commit improves the delta detection used by the CardDAV plugin
when determining the local A/M/R (add/modify/remove) sets, by
ignoring PHOT and X-FAVORITE field changes, and TEL+URL subtype
changes during delta detection, as these can be supported differently
depending on the CardDAV service.

Note that any local change to these details/fields WILL be upsynced
but only if a different change also occurs to the contact (so that
the contact will be detected as "changed").

Contributes to MER#941