[buteo-sync-plugin-carddav] Proceed if MethodNotAllowed encountered on upsync.  Contributes to MER#1304
This commit ensures that we continue the sync even in the case where
the server returns a HTTP 405 MethodNotAllowed error on upsync.
In that case, the server has a Read-Only (or Modify-Only) policy
for the given collection.  We should not fail the entire sync,
but some conflict resolution will occur on next sync.

Contributes to MER#1304